fix: resolve merge conflict
resets UI to state where buy modal was still enabled
